+++ application/controllers/Doctor.php
@@ -31,7 +31,7 @@ class Doctor extends CI_Controller {
 
 	public function adddata()
 	{
-		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true) {
+		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true && $_SESSION['status'] === "ADMIN") {
             
 			//create the data object
 			$data = new stdClass();
@@ -93,7 +93,7 @@ class Doctor extends CI_Controller {
 
 
 	public function editdata($id){
-		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true) {
+		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true && $_SESSION['status'] === "ADMIN") {
             
 			//create the data object
 			$stddata = new stdClass();
@@ -155,7 +155,7 @@ class Doctor extends CI_Controller {
 	}
 
 	public function viewdata($id){
-		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true) {
+		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true && $_SESSION['status'] === "ADMIN") {
 			$data['query'] = $this->DoctorModel->Read_specific($id)->row();
 			$this->load->view('header');
 			$this->load->view('sidebar/users_active');
@@ -168,7 +168,7 @@ class Doctor extends CI_Controller {
 	}
 
 	public function deletedata($ID){
-		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true) {
+		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true && $_SESSION['status'] === "ADMIN") {
 			$data['doctor_id'] = $ID;
 			$this->DoctorModel->Delete($data);
 			$this->DoctorModel->Redirect();
