+++ application/controllers/MedicineType.php
@@ -31,48 +31,56 @@ class MedicineType extends CI_Controller {
 public function adddata()
 	{
 		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true) {
-            
-			//create the data object
-			$data = new stdClass();
 
-			// set validation rules
-			$this->form_validation->set_rules('type_id', 'Medicine Type ID', 'trim|required|alpha_dash|max_length[11]|is_unique[medicine_type.type_id]', array('is_unique' => 'This id already exists. Please choose another one.'));
-			$this->form_validation->set_rules('name', 'Medicine Type Name');
-			
-			if ($this->form_validation->run() === false) {
-			
-				// validation not ok, send validation errors to the view
-				$this->load->view('header');
-            	$this->load->view('sidebar/management_active');
-            	$this->load->view('navbar');
-            	$this->load->view('medicine_type/medicine_type_add_view');
-            	$this->load->view('footer/footer');
-			
-			} else {
-				// set variables from the form
-				$type_id    = $this->input->post('type_id');
-				$name = $this->input->post('name');	
+			if($_SESSION['status'] === "ADMIN" or $_SESSION['status'] === "PHARMACIST"){
+				//create the data object
+				$data = new stdClass();
 
-				if ($this->MedicineTypeModel->create_medicine($type_id,$name)) {
+				// set validation rules
+				$this->form_validation->set_rules('type_id', 'Medicine Type ID', 'trim|required|alpha_dash|max_length[11]|is_unique[medicine_type.type_id]', array('is_unique' => 'This id already exists. Please choose another one.'));
+				$this->form_validation->set_rules('name', 'Medicine Type Name');
 				
-					// user creation ok
-					$this->MedicineTypeModel->Redirect();
-					
-				} else {
+				if ($this->form_validation->run() === false) {
 				
-					// user creation failed, this should never happen
-					$data->error = 'There was a problem creating your new data. Please try again.';
-					
-					// send error to the view
+					// validation not ok, send validation errors to the view
 					$this->load->view('header');
-            		$this->load->view('sidebar/management_active');
-            		$this->load->view('navbar');
-            		$this->load->view('medicine_type/medicine_type_add_view',$data);
-            		$this->load->view('footer/footer');
-					
-				}
+					$this->load->view('sidebar/management_active');
+					$this->load->view('navbar');
+					$this->load->view('medicine_type/medicine_type_add_view');
+					$this->load->view('footer/footer');
+				
+				} else {
+					// set variables from the form
+					$type_id    = $this->input->post('type_id');
+					$name = $this->input->post('name');	
 
+					if ($this->MedicineTypeModel->create_medicine($type_id,$name)) {
+					
+						// user creation ok
+						$this->MedicineTypeModel->Redirect();
+						
+					} else {
+					
+						// user creation failed, this should never happen
+						$data->error = 'There was a problem creating your new data. Please try again.';
+						
+						// send error to the view
+						$this->load->view('header');
+						$this->load->view('sidebar/management_active');
+						$this->load->view('navbar');
+						$this->load->view('medicine_type/medicine_type_add_view',$data);
+						$this->load->view('footer/footer');
+						
+					}
+
+				}
+			} else {
+				
+				redirect('/');
+				
 			}
+            
+			
 
         } else {
             redirect('/');
@@ -82,48 +90,56 @@ public function adddata()
 	public function editdata($id){
 		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true) {
 
-			//create the data object
-			$stddata = new stdClass();
+			if($_SESSION['status'] === "ADMIN" or $_SESSION['status'] === "PHARMACIST"){
+				//create the data object
+				$stddata = new stdClass();
 
-			// set validation rules
-			$this->form_validation->set_rules('type_id', 'Medicine Type ID', 'trim|required|alpha_dash|max_length[11]', array('is_unique' => 'This id already exists. Please choose another one.'));
-			$this->form_validation->set_rules('name', 'Medicine Type Name');
-			
-			if ($this->form_validation->run() === false) {
-			
-				// validation not ok, send validation errors to the view
-				$data['query'] = $this->MedicineTypeModel->Read_specific($id)->row();
-				$this->load->view('header');
-    			$this->load->view('sidebar/management_active');
-        		$this->load->view('navbar');
-				$this->load->view('medicine_type/medicine_type_edit_view',$data);
-				$this->load->view('footer/footer');
-			
-			} else {
-				// set variables from the form
-				$type_id    = $this->input->post('type_id');
-				$name = $this->input->post('name');	
+				// set validation rules
+				$this->form_validation->set_rules('type_id', 'Medicine Type ID', 'trim|required|alpha_dash|max_length[11]', array('is_unique' => 'This id already exists. Please choose another one.'));
+				$this->form_validation->set_rules('name', 'Medicine Type Name');
 				
-				if ($this->MedicineTypeModel->Update($type_id,$name)) {
+				if ($this->form_validation->run() === false) {
 				
-					// user creation ok
-					$this->MedicineTypeModel->Redirect();
-					
-				} else {
-				
-					// user creation failed, this should never happen
-					$data->error = 'There was a problem creating your new account. Please try again.';
-					
-					// send error to the view
+					// validation not ok, send validation errors to the view
+					$data['query'] = $this->MedicineTypeModel->Read_specific($id)->row();
 					$this->load->view('header');
-            		$this->load->view('sidebar/management_active');
-            		$this->load->view('navbar');
-            		$this->load->view('medicine_type/medicine_type_edit_view',$data);
-            		$this->load->view('footer/footer');
+					$this->load->view('sidebar/management_active');
+					$this->load->view('navbar');
+					$this->load->view('medicine_type/medicine_type_edit_view',$data);
+					$this->load->view('footer/footer');
+				
+				} else {
+					// set variables from the form
+					$type_id    = $this->input->post('type_id');
+					$name = $this->input->post('name');	
 					
-				}
+					if ($this->MedicineTypeModel->Update($type_id,$name)) {
+					
+						// user creation ok
+						$this->MedicineTypeModel->Redirect();
+						
+					} else {
+					
+						// user creation failed, this should never happen
+						$data->error = 'There was a problem creating your new account. Please try again.';
+						
+						// send error to the view
+						$this->load->view('header');
+						$this->load->view('sidebar/management_active');
+						$this->load->view('navbar');
+						$this->load->view('medicine_type/medicine_type_edit_view',$data);
+						$this->load->view('footer/footer');
+						
+					}
 
+				}
+			} else {
+				
+				redirect('/');
+				
 			}
+
+			
 
         } else {
             redirect('/');
@@ -132,12 +148,20 @@ public function adddata()
 
 	public function viewdata($id){
 		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true) {
-			$data['query'] = $this->MedicineTypeModel->Read_specific($id)->row();
-			$this->load->view('header');
-			$this->load->view('sidebar/users_active');
-			$this->load->view('navbar');
-			$this->load->view('medicine_type/medicine_type_data_view', $data);
-			$this->load->view('footer/footer');
+
+			if($_SESSION['status'] === "ADMIN" or $_SESSION['status'] === "PHARMACIST"){
+				$data['query'] = $this->MedicineTypeModel->Read_specific($id)->row();
+				$this->load->view('header');
+				$this->load->view('sidebar/users_active');
+				$this->load->view('navbar');
+				$this->load->view('medicine_type/medicine_type_data_view', $data);
+				$this->load->view('footer/footer');
+			} else {
+				
+				redirect('/');
+				
+			}
+			
 		} else {
             redirect('/');
         }
@@ -145,9 +169,17 @@ public function adddata()
 
 	public function deletedata($ID){
 		if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] === true) {
-			$data['type_id'] = $ID;
-			$this->MedicineTypeModel->Delete($data);
-			$this->MedicineTypeModel->Redirect();
+
+			if($_SESSION['status'] === "ADMIN" or $_SESSION['status'] === "PHARMACIST"){
+				$data['type_id'] = $ID;
+				$this->MedicineTypeModel->Delete($data);
+				$this->MedicineTypeModel->Redirect();
+			} else {
+				
+				redirect('/');
+				
+			}
+			
 		} else {
             redirect('/');
         }
